// src/types/resumeSchema.ts
export interface ResumeSchema {
// ...
  skills?: {
    /**
     * e.g. Web Development
     */
    name?: string;
    /**
     * e.g. Master
     */
    level?: string;
    /**
     * List some keywords pertaining to this skill
     */
    keywords?: string[];
    [k: string]: unknown;
  }[];
// ...
}